Adding FTL as a streaming backend
