+++ assets/js/app.js
@@ -13,90 +13,34 @@ import "../css/app.scss"
 //     import socket from "./socket"
 //
 import "phoenix_html"
-import { Socket } from "phoenix"
-import { LiveSocket } from "phoenix_live_view"
+import {
+    Socket
+} from "phoenix"
+import {
+    LiveSocket
+} from "phoenix_live_view"
 import BSN from "bootstrap.native";
-
-
-function loadLib(url) {
-    return new Promise((resolve, reject) => {
-        // adding the script tag to the head as suggested before
-        let head = document.getElementsByTagName('head')[0];
-        let script = document.createElement('script');
-        script.type = 'text/javascript';
-        script.src = url;
-
-        // then bind the event to the callback function
-        // there are several events for cross browser compatibility
-        script.onreadystatechange = resolve;
-        script.onload = resolve;
-
-        // fire the loading
-        head.appendChild(script);
-    });
-}
-
-function tryVideo(url) {
-    if (document.getElementById('player')) {
-        Promise.all([
-            loadLib("https://cdnjs.cloudflare.com/ajax/libs/dashjs/2.9.3/dash.all.min.js"),
-            loadLib("https://cdn.jsdelivr.net/npm/hls.js@latest"),
-            loadLib("/ovenplayer/ovenplayer.js"),
-        ]).then(data => {
-            console.log('finished loading video player with ' + url);
-            // OvenPlayer.debug(true);
-
-            let player = OvenPlayer.create("player", {
-                // autoStart: true,
-                image: "/images/stream-not-started.jpg",
-                sources: [
-                    {
-                        type: "webrtc",
-                        file: "wss://edge.live.glimesh.tv:3334/app/stream",
-                        label: "Warp 1"
-                    },
-                    {
-                        type: "hls",
-                        file: "https://edge.live.glimesh.tv/app/stream/playlist.m3u8",
-                        label: "HLS"
-                    },
-                    {
-                        type: "mpd",
-                        file: "https://edge.live.glimesh.tv/app/stream/manifest.mpd",
-                        label: "MPEG-DASH"
-                    }
-                ]
-            });
-            player.on("error", function (error) {
-                console.log(error);
-            });
-
-            player.on('metaChanged', function (f) {
-                player.play();
-            });
-        });
-    }
-}
 
 import ProcessPayment from './hooks/ProcessPayment';
 import Chat from './hooks/Chat';
-import Choices from "./hooks/Choices"
+import Choices from "./hooks/Choices";
+import FtlVideo from "./hooks/FtlVideo";
+import OvenVideo from "./hooks/OvenVideo";
 
 let Hooks = {};
 Hooks.ProcessPayment = ProcessPayment;
 Hooks.Chat = Chat;
 Hooks.Choices = Choices;
-Hooks.LoadVideo = {
-    playbackUrl() {
-        return this.el.dataset.playbackUrl
-    },
-    mounted() {
-        tryVideo(this.playbackUrl())
-    }
-};
+Hooks.FtlVideo = FtlVideo;
+Hooks.OvenVideo = OvenVideo;
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content");
-let liveSocket = new LiveSocket("/live", Socket, { params: { _csrf_token: csrfToken }, hooks: Hooks });
+let liveSocket = new LiveSocket("/live", Socket, {
+    params: {
+        _csrf_token: csrfToken
+    },
+    hooks: Hooks
+});
 
 // Show progress bar on live navigation and form submits
 window.addEventListener("phx:page-loading-start", () => {});
@@ -116,4 +60,3 @@ liveSocket.connect();
 // liveSocket.enableDebug();
 // liveSocket.enableLatencySim(1000)
 window.liveSocket = liveSocket;
-
