+++ assets/package.json
@@ -11,6 +11,7 @@
     "OvenPlayer": "github:AirenSoft/OvenPlayer",
     "bootstrap.native": "^3.0.9",
     "choices.js": "^9.0.1",
+    "janus-ftl-player": "0.0.2-alpha",
     "phoenix": "file:../deps/phoenix",
     "phoenix_html": "file:../deps/phoenix_html",
     "phoenix_live_view": "file:../deps/phoenix_live_view"
